Refactored re-rendering chain
1. Moved LoadFavouritesIds and SaveFavouritesIds to new file: StorageActions.js.
2. Added some new comments.
3. Added filtering pokemons to re-rendering chain.
